adjust count

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 
 size_t print_list(const list_t *h)
 {
-	unsigned int index = 0;
+	size_t index = 0;
 	unsigned int count = 0;
 
 	while (h != NULL)
